Improve CI Tests
Switches the disk layout configuration based on the hypervisor type
which selects the HBA type volumes are attached via and their physical
layout as far as SCSI is concerned.  Removes the hard coded sleep which
is unreliable and replaces with a repeating netcat until the port comes
up.  Fails after 5 minutes