Support for Centos 7
Support for RedHat based systems.  Abstracts away differences between Ubuntu
and Centos package names and service providers, which are required for init
and upstart to work correctly with monitors and gateways.  Service invokation
is switched based upon the OS too.  The OSD type is updated to activate disks
on Centos (this happens implicitly via upstart on Ubuntu).  Default parameters
have been updated to use the loopback device, this means we aren't reliant on
a VM having fixed interface naming (e.g. eth0) for CI.  Tests added for local
and cloud Centos.  Acceptance tests updated to use netstat to check for a
working rados gateway as the previous nc command doesn't exist on Centos.