Fix Services
Centos services were not cleanly restarted after a reboot.  The service providers
have been changed to upstart or systemd.  For upstart the 'upstart' file required
by the init scripts has been rolled into the case before the service is started.
For systemd the service is enabled before starting.  In both cases it sadly made
for nicer code to start the service via an exec.  A new service class ensures the
top level ceph.target service is enabled on systemd to trigger the start of all
dependant services.  This reboot functionality is protected via extended CI tests.