Support Jewel and Ubuntu Xenial
Fixes to the OSD provider as /dev/disk/by-parttypeuuid has been dropped by Ceph
so now just call sgdisk on the partition to see if it's an OSD type.  Support
has been added to params to check if we are running Xenial which now uses
systemd in place of upstart (yey: parity).  The apt repository is now installed
based on the operating system code name.  New CI tests have been added for Puppet
4 on Xenial to test for regressions.  Finally using localhost doesn't work on
Jewel anymore so we probe Beaker and get the host IP address and use that in
the dynamically generated configuration.