Genericise OSD Type
The osd type and provider have been updated to be more extensible.  The journal
parameter is now a parameter of its own now separated from the resource name.  It
is optional now.  The fstype parameter has now been dropped in favour of a params
hash which may contain arbitrary key value pairs to be passed to ceph-disk, enabling
the possibility of bluestore and dmcrypt with minimal changes.

The API changes are not backwards compatible and will need manifests/hiera updating.
Due to the nature of the change it will not break existing clusters or attempt to
make changes.