Basic CephFS Support
Provisions mds specific directory structures and spawns the necessary services.  For
both mds and rgw services the logic has been modified to use bootstrap keys to generate
client keys for each of the services dynamically.  The biggest effect of this change
is on rados gateways, in particular prior installations that named gateways client.radosgw
must be updated to use client.rgw.  The upgrade path will be documented in the 1.6 release
notes