Force Systemctl Reload
On multiple occasions now when provisioning OSDs I've seen devices not have the
ceph:ceph udev rules applied and things got wuite wrong.  The online fix appears
to be a reload of systemctl, so force this once we've installed and enabled the
ceph target.  This is a strictly pre-emptive action, with no guarantees of working