Fix Undef OSD Parameter Type Checking
Version 1.2.0 introduced a bug where a parameter to be passed to ceph-disk without
a value would fail as the Ceph::Disks struct only tests for a String.  This has
been updated to an optional string.  Tests guard this behaviour, additionally
journal size has been changed to an int to check that code path also.