Improve RGW Type Checking
When an RGW ID is supplied as a parameter it is only checked whether or not it
is a string.  See 42e4d13cb7fb4a2b9ed4d34a1fee2222cf9988f8 for a more detailed
explanation.  A new data type is introduced to sanitize the rgw_id parameter
and ensure it is in the correct form.